+++ Tests/Datagrid/ProxyQueryTest.php
@@ -11,15 +11,26 @@
 
 namespace Sonata\DoctrineORMAdminBundle\Tests\Datagrid;
 
+use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\Query\Expr\From;
+use Sonata\DoctrineORMAdminBundle\Tests\Fixtures\DoctrineType\UuidType;
+use Sonata\DoctrineORMAdminBundle\Tests\Fixtures\Util\NonIntegerIdentifierTestClass;
 
 class ProxyQueryTest extends \PHPUnit_Framework_TestCase
 {
+    public static function setUpBeforeClass()
+    {
+        if (!Type::hasType('uuid')) {
+            Type::addType('uuid', 'Sonata\DoctrineORMAdminBundle\Tests\Fixtures\DoctrineType\UuidType');
+        }
+    }
+
     public function dataGetFixedQueryBuilder()
     {
         return array(
-            array('aaa', 'bbb', 'id', 'id_idx', 33),
-            array('aaa', 'bbb', 'id.value', 'id_value_idx', 33),
+            array('aaa', 'bbb', 'id', 'id_idx', 33, Type::INTEGER),
+            array('aaa', 'bbb', 'id.value', 'id_value_idx', 33, Type::INTEGER),
+            array('aaa', 'bbb', 'id.uuid', 'id_uuid_idx', new NonIntegerIdentifierTestClass('80fb6f91-bba1-4d35-b3d4-e06b24494e85'), UuidType::NAME),
         );
     }
 
@@ -30,7 +41,7 @@ class ProxyQueryTest extends \PHPUnit_Framework_TestCase
      * @param $alias
      * @param $id
      */
-    public function testGetFixedQueryBuilder($class, $alias, $id, $expectedId, $value)
+    public function testGetFixedQueryBuilder($class, $alias, $id, $expectedId, $value, $identifierType)
     {
         $meta = $this->getMockBuilder('Doctrine\ORM\Mapping\ClassMetadataInfo')
             ->disableOriginalConstructor()
@@ -38,6 +49,9 @@ class ProxyQueryTest extends \PHPUnit_Framework_TestCase
         $meta->expects($this->any())
             ->method('getIdentifierFieldNames')
             ->willReturn(array($id));
+        $meta->expects($this->any())
+            ->method('getTypeOfField')
+            ->willReturn($identifierType);
 
         $mf = $this->getMockBuilder('Doctrine\ORM\Mapping\ClassMetadataFactory')
             ->disableOriginalConstructor()
@@ -47,12 +61,26 @@ class ProxyQueryTest extends \PHPUnit_Framework_TestCase
             ->with($this->equalTo($class))
             ->willReturn($meta);
 
+        $platform = $this->getMockBuilder('Doctrine\DBAL\Platforms\PostgreSqlPlatform')
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $conn = $this->getMockBuilder('Doctrine\DBAL\Connection')
+            ->disableOriginalConstructor()
+            ->getMock();
+        $conn->expects($this->any())
+            ->method('getDatabasePlatform')
+            ->willReturn($platform);
+
         $em = $this->getMockBuilder('Doctrine\ORM\EntityManager')
             ->disableOriginalConstructor()
             ->getMock();
         $em->expects($this->any())
             ->method('getMetadataFactory')
             ->willReturn($mf);
+        $em->expects($this->any())
+            ->method('getConnection')
+            ->willReturn($conn);
 
         $q = $this->getMock('PDOStatement');
         $q->expects($this->any())